# openapi.yml
openapi: 3.0.0
info:
  description: Backend API for the Stanford digital repository
  version: 1.0.0
  title: DOR Services API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://dor-services-{env}.stanford.edu"
    description: Production service
    variables:
      env:
        default: prod
  - url: "https://dor-services-{env}.stanford.edu"
    description: Staging service
    variables:
      env:
        default: stage
...
paths:
...
  "/v1/objects/{object_id}/versions/current":
    get:
      tags:
        - versions
      summary: Return the current version number for this object
      description: ""
      operationId: "versions#current"
      responses:
        "200":
          description: OK
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
    delete:
      tags:
        - versions
      summary: Destroys current open version for this object
      description: ""
      operationId: "versions#destroy_current"
      responses:
        "204":
          description: Deleted
        "409":
          description: Current version is not open
        "404":
          description: Object not found
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
...
  "/v1/objects/{object_id}/versions/current/close":
    post:
      tags:
        - versions
      summary: Close the currently open version for this object
      description: ""
      operationId: "versions#close_current"
      responses:
        "200":
          description: OK
      parameters:
        - name: object_id
          in: path
          description: ID of object
          required: true
          schema:
            $ref: "#/components/schemas/Druid"
        - name: start_accession
          in: query
          schema:
            type: boolean
            default: true
          required: false
          description: Indicates if accessionWF should be started after closing the version
        - name: description
          in: query
          description: the description of the version change
          required: false
          schema:
            type: string
            example: "Re-accessioning this object"
        - name: user_name
          in: query
          description: the username of the person creating the version
          schema:
            type: string
            example: "some_sunetid"
        - name: user_versions
          in: query
          description: Create, update, or do nothing with user versions on close
          schema:
            type: string
            enum:
              - "none"
              - "new"
              - "update"
              - "update_if_existing"
...
  /v1/objects/versions/status:
    post:
      tags:
        - versions
      summary: Finds versions for multiple objects.
      description: |-
        This endpoint is used to find the versions for multiple objects. The request body should be an array of druids.

        The response will be an object with the druids as keys and the version status as values.

        If an object is not found, it will not be included in the response.
      operationId: "versions#batch_status"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              # Using an object instead of array due to https://github.com/interagent/committee/issues/220
              type: object
              properties:
                externalIdentifiers:
                  type: array
                  items:
                    $ref: "#/components/schemas/Druid"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/VersionStatus"
...
components:
  schemas:
...
    Druid:
      type: string
      pattern: "^druid:[b-df-hjkmnp-tv-z]{2}[0-9]{3}[b-df-hjkmnp-tv-z]{2}[0-9]{4}$"
      example: "druid:bc123df4567"
...
    VersionStatus:
      type: object
      required:
        - versionId
        - open
        - openable
        - assembling
        - accessioning
        - closeable
        - discardable
        - versionDescription
      properties:
        versionId:
          type: integer
          description: the number of the version
        open:
          type: boolean
          description: whether the version is open
        openable:
          type: boolean
          description: whether the version can be opened
        assembling:
          type: boolean
          description: whether there is an active assembly workflow (not just assemblyWF)
        accessioning:
          type: boolean
          description: whether there is an active accessioningWF workflow
        closeable:
          type: boolean
          description: whether the version can be closed
        versionDescription:
          type: string
          description: a message that explains what changed
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    Error:
      type: object
      properties:
        title:
          type: string
          description: "a short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem."
          example: Invalid Attribute
        detail:
          type: string
          description: a human-readable explanation specific to this occurrence of the problem.
          example: Title must contain at least three characters.
        meta:
          type: object
          description: used to include non-standard meta-information
          example: { backtrace: ["where things went wrong"] }
        source:
          type: object
          properties:
            pointer:
              type: string
              example: /data/attributes/title
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []